requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
tortoise-orm>=0.20.0
aerich>=0.7.2
aiomysql>=0.2.0
pydantic>=2.5.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 